Refactored codebase into separate objects
